% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{catalog}
\alias{catalog}
\title{catalog}
\usage{
catalog(route = 0)
}
\arguments{
\item{route}{Bucket object. String of the bucket object in S3.}
}
\description{
Brings a catalog stored in S3 into a dataframe. By default the
"catalogo de beneficios"
}
\examples{
catalogo <- catalog()
}
