% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{clear_results}
\alias{clear_results}
\title{clear_results}
\usage{
clear_results(connection)
}
\arguments{
\item{connection}{DBI connection. A connection to a database must be open and given.}
}
\description{
This function clears the results from a previous executed
query.
}
\examples{
clear_results(con)
}
