% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{:=}
\alias{:=}
\title{multireturn}
\usage{
":="(lhs, rhs)
}
\description{
Allows to return and save multiple variables within a function
}
\examples{
c(programas_persona,los_queries) := load_or_run(con,query,los_queries)
}
