% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrive.R
\name{cross_tables}
\alias{cross_tables}
\title{cross_tables}
\usage{
cross_tables(table_1, table_2, key_1, key_2)
}
\arguments{
\item{table_1}{a tbl-like object}

\item{table_2}{a tbl-like object}

\item{key_1}{the column name from table_1 to compare}

\item{key_2}{the column name from table_2 to compare}
}
\description{
Returns a match between two tbl by defined key
}
\examples{
cross_tables(domicilios_sample_query,cuis_sample,llave_hogar_h,llave_hogar_h)
}
