% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{csv_s3}
\alias{csv_s3}
\title{csv_s3}
\usage{
csv_s3(object = "s3://pub-raw/diccionarios/catalogo_beneficio.csv")
}
\arguments{
\item{route}{Bucket object. String of the bucket object in S3.}
}
\description{
Brings a CSV stored in S3 into a dataframe. By default the
"catalogo de beneficios"
}
\examples{
catalogo <- csv_s3()
}
