% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuis_raw_sample.R
\name{cuis_raw_sample}
\alias{cuis_raw_sample}
\title{Sample CUIS raw table}
\usage{
cuis_raw_sample(con, p = 0.01, seed = 1234)
}
\arguments{
\item{con}{a DBIConnection object to Postgres predictiva data base, as
returned by \code{\link[DBI]{dbConnect}}.}

\item{p}{Sample size as proportion of the total number of households
in CUIS, defaults to 0.01.}

\item{seed}{numeric value to be used to generate a seeding for the PRNG
random generator in Postgres backend.}
}
\description{
Random Sample of CUIS table on the raw schema.
}
