% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{discon_db}
\alias{discon_db}
\title{discon_db}
\usage{
discon_db(connection)
}
\arguments{
\item{connection}{DBI connection. A connection to a database must be open and given.}
}
\description{
This function disconnects a PostgreSQL
connection.
}
\examples{
discon_db(con)
}
