% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{large_table}
\alias{large_table}
\title{large_table}
\usage{
large_table(connection, schema, the_table)
}
\arguments{
\item{connection}{DBI connection. A connection to a database
must be open and given.}

\item{schema}{variable. A valid schema from a database on the
connected database.}

\item{the_table.}{An existing table in the given schema.}
}
\description{
This function loads a connection to a large table without
loading it to memory.
}
\examples{
cuis_table <- large_table(con,raw,cuis_39_9)
}
