% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{load_geom}
\alias{load_geom}
\title{load_geom}
\usage{
load_geom(connection, schema, the_table,
  columns = "cve_mun, cve_ent, cve_muni, ", geom_col, col_shape,
  options = "")
}
\arguments{
\item{route}{Bucket object. String of the bucket object in S3.}
}
\description{
Gives a "ready to go" data frame for geometry plotting
}
