% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrive.R
\name{load_or_run}
\alias{load_or_run}
\title{load_or_run}
\usage{
load_or_run(connection, query, the_dic)
}
\arguments{
\item{connection}{An Athena connection}

\item{query}{charachter. A query string}

\item{the_dic}{The previously loaded query dictionary}
}
\description{
Checks if a query has already been runned in Athena and if it has then
loads the previous results and if it hasn't then it runs it and saves the
query for further usage
}
