% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{load_query}
\alias{load_query}
\title{load_query}
\usage{
load_query(connection, schema, the_table, columns = "*", options = "")
}
\arguments{
\item{connection}{DBI connection. A connection to a database}

\item{schema}{variable. A valid schema from a database on the}

\item{options}{string. Part of the SQL query with containing WHERE, ORDER,
LIMIT and so statements}

\item{the_table.}{An existing table in the given schema.}

\item{colums}{string. The columns in the database we want to retrieve
information}
}
\description{
Gives a "ready to go" data frame for geometry plotting
}
