% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{mun_geom}
\alias{mun_geom}
\title{mun_geom}
\usage{
mun_geom(connection, ent = "")
}
\arguments{
\item{connection}{DBI connection. A connection to a database}

\item{ent}{char. The string corresponding to the cveent, if blank returns
national entities.}
}
\description{
Loads the municipalities geometries nation-wide or entity-wide
}
\examples{
geom_national <- mun_geom(con1)
geom_tlaxcala <- mun_geom(con1,"29")
}
