% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{prev_connect}
\alias{prev_connect}
\title{prev_connect}
\usage{
prev_connect()
}
\arguments{
\item{No}{parameter is needed.}
}
\description{
This function takes the Postgres environment variables
to automatically connect to the preventivadb database.
The connection must be saved into a variable in order to
use future functions.
}
\examples{
con <- prev_connect()

}
