% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{pub_connect}
\alias{pub_connect}
\title{pub_connect}
\usage{
pub_connect(s3dir, schema)
}
\arguments{
\item{s3dir}{The S3 route to save the query results}

\item{schema}{The existing schema in Athena}
}
\description{
This function connects to an Athena schema and sets the saving
directory in AWS' S3.
to automatically connect to the preventivadb database.
The connection must be saved into a variable in order to
use future functions.
}
\examples{
con <- pub_connect()
}
