% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout.R
\name{quant_labels}
\alias{quant_labels}
\title{quant_labels}
\usage{
quant_labels(variable, no_classes = 6)
}
\arguments{
\item{variable}{data.frame. Te data we need to divide}

\item{no_classes}{Int. The number of quantile divisions.}
}
\description{
Make quantile divisions of a data to plot

A theme for plots in Rmd
}
\examples{
municipios_nal$cuenta_q <- quant_labels(municipios_nal$cuenta)
}
