% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_result}
\alias{retrieve_result}
\title{retrieve_result}
\usage{
retrieve_result(query, n = -1, number = Inf)
}
\arguments{
\item{query}{An exec unfetched query}
}
\description{
Return the fetch results of a query
}
\examples{
sample_table(load_table(prev_connect(),raw,sifode))
}
