% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrive.R
\name{retrive_result}
\alias{retrive_result}
\title{retrive_result}
\usage{
retrive_result(query)
}
\arguments{
\item{query}{An exec unfetched query}
}
\description{
Return the fetch results of a query
}
\examples{
sample_table(load_table(prev_connect(),raw,sifode))
}
