% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{sample_table}
\alias{sample_table}
\title{sample_table}
\usage{
sample_table(connection, p = 0.01, seed = 1234, schema, the_table,
  columns = "*", options = "", lim = 0)
}
\arguments{
\item{connection}{DBIConnection object to Postgres predictiva data base, as
returned by \code{\link[DBI]{dbConnect}}.}

\item{p}{Sample size as proportion of the total number of rows in a
table, defaults to 0.01.}

\item{seed}{numeric value to be used to generate a seeding for the PRNG
random generator in Postgres backend.}

\item{options}{string. Part of the SQL query with containing WHERE, ORDER,
LIMIT and so statements}

\item{lim}{int. A limit for SQL lines to return.}

\item{colums}{string. The columns in the database we want to retrieve
information defaults to all}
}
\description{
Random Sample of any given table and schema.
}
\examples{
sample_table(con,0.01,1234,raw,cuis_historico_domicilios)
}
