% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_layout.R
\name{theme_pub}
\alias{theme_pub}
\title{theme_pub}
\usage{
theme_pub(base_size = 12, font = NA)
}
\description{
A theme for plots in Rmd
}
\examples{
ggplot(data=municipios_tlax, aes(x=municipio, y=cuenta)) +
 geom_bar(stat="identity") +
 labs(title="Número de beneficiarios por municipio", fill="Número de beneficiarios",y="Beneficiarios únicos") +
 scale_y_continuous(labels = comma,breaks = pretty_breaks(n = 6)) +
 theme_pub()
}
