% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{write_s3}
\alias{write_s3}
\title{write_s3}
\usage{
write_s3(dataf, name, s3bucket = Sys.getenv("S3_DIR"))
}
\arguments{
\item{dataf}{The data.frame objet to write as a CSV}

\item{name}{The name of the file in the S3 bucket}

\item{s3bucket}{The name of the S3 bucket}
}
\description{
Writes a dataframe as a CSV in a S3 bucket
}
\examples{
write_s3(the_dic, "dict/fun_dict.csv", Sys.getenv("S3_DIR"))
}
