% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{cuis_raw_sample}
\alias{cuis_raw_sample}
\title{sample_table}
\usage{
cuis_raw_sample(connection, p = 0.01, seed = 1234)
}
\arguments{
\item{connection}{DBIConnection object to Postgres predictiva data base, as
returned by \code{\link[DBI]{dbConnect}}.}

\item{p}{Sample size as proportion of the total number of households
in CUIS, defaults to 0.01.}

\item{seed}{numeric value to be used to generate a seeding for the PRNG
random generator in Postgres backend.}
}
\description{
Random Sample of any given table.
}
