% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrive.R
\name{join_tables}
\alias{join_tables}
\title{join_tables}
\usage{
join_tables(left_table, right_table, left_key, right_key)
}
\arguments{
\item{left_table}{a tbl-like object}

\item{right_table}{a tbl-like object}

\item{left_key}{the column name from left_table to compare}

\item{right_key}{the column name from right_table to compare}
}
\description{
Returns a match between two tbl by defined key
}
\examples{
cross_tables(domicilios_sample_query,cuis_sample,llave_hogar_h,llave_hogar_h)
}
