% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{load_query}
\alias{load_query}
\title{load_query}
\usage{
load_query(connection, schema, the_table, columns = "*", options = "")
}
\arguments{
\item{connection}{DBI connection. A connection to a database
must be open and given.}

\item{schema}{variable. A valid schema from a database on the
connected database.}

\item{the_table.}{An existing table in the given schema.}

\item{the_columns.}{The columns to retrieve}
}
\description{
This function loads a connection to
a given table from a particular schema in a database
connection.
}
\examples{
the_dic<-load_table(con,raw,sifode_dic)
}
